college/requirements.txt
streamlit
pandas
numpy
mysql-connector-python
plotly
matplotlib
seaborn
